Added more flexibility in setting options
Now there are three ways to set options framework options:

  1. The old optionsframework_options() method
  2. Returning an array() in options.php
  3. Adding a filter on 'of_options'

The advantage of #3 is that this provides more granular control
over options.  For example, a child theme could add its own options
in addition to the parent theme's options using this method.  It
also means that options can be set anywhere in your project--the
options.php file isn't necessarily required (well, it still needs
to be there for optionsframework_option_name(), but I plan to
change that, too, soon).